requirements.txt
griptape[google]
ollama
streamlit
python-dotenv
